Corrigindo erros para Test com JUnit

+++ blogpessoal/src/main/java/br/com/generation/blogpessoal/controller/PostagemController.java
@@ -1,6 +1,7 @@
 package br.com.generation.blogpessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -14,6 +15,7 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import br.com.generation.blogpessoal.model.Postagem;
 import br.com.generation.blogpessoal.repository.PostagemRepository;
@@ -55,6 +57,12 @@ public class PostagemController {
 	
 	@DeleteMapping("/{id}")
 	public void deletePostagem(@PathVariable long id) {
-		postagemRepository.deleteById(id);
+			Optional<Postagem> postagem = postagemRepository.findById(id);
+		
+			if (postagem.isPresent()) {
+				postagemRepository.deleteById(id);
+			} else {
+				throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Não foi possível deletar, pois o id que você digitou não existe!");
+		}
 	}
 }
